+++ CHANGES.md
@@ -5,7 +5,7 @@
 
 - Renamed `magic` command to `plot`, and `plot` to `flexplot`.
 - `plot`: new default metric `proc_mem_rss_percent`, do not plot `system_loadavg1` anymore by default.
-- `plot`: improve subtitle
+- `plot`: improved subtitle
 
 
 ## 0.2.0 (2019-08-13)